Add lang field to GradedForm
and add a special case for hidden fields in graded_form.html.
Get language from POST if available.

This handles the language mismatch bug in every case that applies to O1,
except for the piazza exercise. Because whether A+ loads the previous submission
is decided by the course material and not grader, it is almost impossible to
fix the bug in general.