Modify grading of checkbox questions in questionnaires (#42)
Add partial points option to the checkbox questions. When set,
the student is rewarded with some points for partially correct
answers. Choosing half of the correct choises grants 0 points and
points increase linearly when more correct choices are made.

Change the grading semantics when no correct options are set
in a checkbox question. This is a backwards incompatible change,
but it should not cause issues since real exercises are expected
to define correct options. Previously, when no correct options
were defined, any answer was granted points for the question.
Now, only an empty answer receives points. The idea is that
if there are no correct options, then the correct answer is to
select nothing. Feedback and enrollment questionnaires are
handled differently so that they always grant full points
even if no correct options are set.

Add neutral options to checkbox questions. Neutral options do not
affect the grading. The student may select them or leave them
unselected.

Feedback and enrollment questionnaires always grant full points.
The correct answers need not be defined in the configuration.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>