Add the 'required' HTML attribute to required form fields (#45)
If a questionnaire field or a file field in an acceptFiles exercise
is set as required in the YAML configurations, the rendered
HTML of the field uses the required HTML attribute.
(The attribute is not used with multiple checkboxes since then
the user would have to select every checkbox.)
The HTML attribute causes web browsers to check that the user
has input something to the field.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>