Validate numeric inputs in questionnaire text fields (#43)
The compare_method setting of questionnaire text fields is used
to recognize numeric fields (int and float). The grader now
validates that the user input is numeric in the numeric text
fields. Non-numeric input is rejected (i.e., the submission
receives status "rejected").

Additionally, the error tolerance of the float fields is now
relational rather than absolute. An error of 2% is accepted
relative to the larger value (larger of the two: user input and
the model solution).

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>