Radio button questions require that correct options are set
Change the grading semantics when no correct options are set
in a radio button question. This is a backwards incompatible change,
but it should not cause issues since real exercises are expected
to define correct options. Previously, when no correct options
were defined, any answer was granted points for the question.
Now, all answers are incorrect if no correct options have been
configured.

Feedback and enrollment questionnaires are handled differently
so that they always grant full points even if no correct options
are set.

Similar change was made for checkbox questions in the commit
https://github.com/Aalto-LeTech/mooc-grader/commit/685b556298ad5cdc0f902611346611ab48531b85
(that commit had other changes too).