Add randomized checkbox questions and fix pick_randomly (#44)
Add randomized and correct_count options to checkbox questions.
Checkbox questions can randomly select a subset of the answer
choices that are shown to the user at a time. The random
selection changes after each submission, but it does not
change if the user simply reloads the web page.

Fix issues with the pick_randomly option of questionnaires.
The randomly selected questions no longer change when the user
reloads the page.

Questions are numerated automatically when the question title
contains the substring {#}. This is useful for the pick_randomly
option since the order of the questions is not known beforehand.