Fix issues in random questions and update README (#44)
MOOC-grader gives a clear rejected feedback to A+ when the random
checksums are invalid in the submission. Previously, A+ showed
a generic error even though there is no error in the grader.

Used nonces are saved in random checkbox questions. This is
needed in order to prevent users from reusing previous nonces.
Reusing old nonces and checksums would enable users to answer
old questions even though the questions are supposed to change
randomly.

The current_sample function used to randomly select questions
now uses a separate random number generator so that the
deterministically set seed does not interfere with other uses
of the random module.

The question form is now always disabled in the feedback for
random checkbox questions. Previously, it was sometimes enabled
even though the submit button was not rendered.

Random checkbox questions do not generate a new nonce value for
the feedback HTML since the feedback form is disabled and can not
be submitted. The nonce and checksum are not included in the
feedback HTML at all when the form is accepted for grading.

current_sample: the randomly generated sample of question indexes
is not sorted so that the questions can be shown in any random
order. The correct choices in random checkbox questions are not
always listed before the incorrect choices; instead, the order
of the choices is random.

The seed is now based on the exercise and question keys in
addition to the user IDs and the submission ordinal number.
Thus, if the the same exercise is defined multiple times in the
chapter, the random choice of questions is different for each
exercise. Previously, each exercise made the same identical
random sample.

Nonces are generated deterministically for random questionnaires.
If the exercise is reloaded, then the nonce does not change.
The randomly selected questions do not change either.
If the nonce changed, then the student could gather multiple nonces
for the same set of randomly selected questions and answer those
questions repeatedly with different nonces. That is prevented by
not generating new nonces and checksums for the same questions.