Questionnaire feedback does not reveal the model solution (#48)
Questionnaire feedback used to reveal the model solution when
the student had consumed all of the submission attempts.
This commit adds a new setting 'reveal_model_at_max_submissions'
that controls whether the model solution is revealed in the
feedback. It is false by default because many teachers were not
aware that the questionnaire feedback could reveal model
solutions even before the exercise deadline.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>