Remove nonces from random questionnaires
Nonces are not needed because the grader now computes the random sample of
questions deterministically in the server. Since the grader does not use
the sample data from the submission data anymore, there is no need to protect
against replay attacks and the nonces are unnecessary.

The sample is still stored in the hidden HTML form field and the submission
data for debugging purposes.

If a random questionnaire is modified after students have submitted to it,
the grader could compute different random samples after the modification
because the random seed depends on the number of available questions.
It is recommended that random questionnaires are not modified after releasing
them to the students.