Add regexp-based feedback and subdiff method to text questions (#49)
Feedback can be given with regular expressions in questionnaire text questions.
If the submission matches the regexp, the feedback is shown. There may be
multiple lines of feedback for different submissions. Regexp support must be
enabled separately with the configuration field `compare_regexp` under the
`feedback` object of the exercise YAML file. By default, string equality is
used to compare the values without any regexp matching.

The subdiff compare method supports multiple correct answer choices.
The choices are compared to the submission with string equality (not regexp).
In addition to the normal feedback defined in the configuration, subdiff
questions generate automatic feedback that shows the difference of the
student's submission to the correct answers. The difference feedback reveals
exactly which characters are correct in the student's submission.