Add show_model_answer setting to questionnaires (#50)
The setting `show_model_answer` in the exercise YAML can be used
to disable model solutions for questionnaire exercises in A+.
If it is set to False, A+ does not show the model solution to
students after the module deadline. (This refers to the
"show model answer" button in the A+ exercise navigation bar.)

By default, the setting is True and A+ shows the model answer
after the deadline.