Show indication of wrong or right answer in question feedback
- v and x -signs have been added to hints displaying at the end
of each question.
- Texts "Correct!" and "Incorrect" have been added to each question
to add the feedback for the student and to not make
the questionnaire feedback rely so heavily on only color.
- Chosen options have been highlighted
- After the student has reached max-submissions, if the answer is not
correct, but staff has set the "reveal-model-at-max-submissions" to be true,
correct answers are shown alongside the fields.
- These html-elements with "data-aplus-once" are only to be shown
when the data is first stored to the database. Thus correct answers
cannot be checked later on, only right after the max submissions have
been reached. This has not been changed from the previous implementation.