Remove the unnecessary questionnaire grading feedback
There was a alert on top of the questionnaire showing the
points that the student got. This removes that alert, as the
feedback of grading has been otherwise improved.

Alternatively the upper part showed the "accepted_message"
that the teacher had set to be shown when an asynchronous submission
was accepted. This would have been shown regardless whether the
student got full points of the questionnaire. This logic has been
preserved, so that teacher can still set accepted_messages, if they
wish.