Clarify the way correct answers are shown in questionnaires
There are two types of model answers for questionnaires, the
ones shown in a separate modal if "show-model-answer" is set to true.
And the ones shown within the form itself at the point of the max
submissions, if the "reveal-model-at-max-submissions" is set to true.
This clarifies their role and visibility within the form and hides
hints if the created form is a model answer.