Change the title "Submission accepted" to "Submission received"
The async_accepted.html template is used when a submission is
submitted for asynchronous grading and the grading is going to
be started.

The message "submission accepted" was a little misleading in
the sense that it may make the user think that the submission
is graded with acceptable points, but now it only means that
the submission has been received and will be graded soon.

In addition, add a comment to the message that the user may close
the modal dialog. The user need not wait and stare at the modal
dialog while the submission is being graded. In some courses,
especially in Scala courses, the grading usually takes over one
minute and the user may think it is necessary to wait a minute
in the modal dialog before they can move on. The modal dialog
polls for the feedback for one minute and then informs the user
that the grading takes longer than that.