Connect question label to the input
For multiple-choice questions, the label of the choice was not connected
to the input element (checkbox or radio button). Because of this,
clicking the label would not activate the input element. This commit
fixes the issue by ensuring that the label's `for` attribute matches the
input element's `id` attribute.

Fixes #107