Fix checkbox hints for Django 2.2
The upgrade to Django 2.2 caused checkbox-based exercises to no longer
display all hints. This is because before the upgrade, iterating over
the BoundField would yield CheckboxChoiceInput instances, and the
graded_form template rendering relied on two attributes that were
available in that class (value and choice_value). In Django 2.2, the
iteration yields BoundWidget instances, which don't have the previously
used attributes. This commit updates the template to use attributes that
are compatible with Django 2.2, and reverts a previous change to
find_common_hints, as it is no longer necessary.

Fixes #110