Ensure unique IDs in exercise forms
Some exercise types (acceptPost, acceptFiles, acceptGeneralForm) used
field names directly as HTML input element IDs. This was problematic
when multiple exercises were displayed on the same page in A+, and some
exercises used some identical field names. This caused duplicate IDs,
which is invalid HTML. This commit fixes the issue be prefixing the IDs
with a random component, e.g. "exercise-XXXXXXXX-field1". This format is
consistent with the one used in GradedForm.

Fixes #111