Don't reveal answers when submission is rejected
When the user is submitting their last available submission to a
questionnaire, the correct answers would get revealed regardless of
whether the submission was accepted. This was problematic because
rejected submissions do not consume the remaining available submissions.
Because of this, the user would have one submission left after having
seen the correct answers, allowing them to get full points easily. This
commit fixes the issue be ensuring that the correct answers are only
revealed if the submission was not rejected.

Fixes #103