Support two Sphinx versions in the gitmanager course build
This is needed for a transitional period when a new Sphinx version
is supported in the production server, but some courses are not
yet ready to use it. Such courses would keep using the old Sphinx
version until they are ready to switch.

This change is going to be deployed with the new Kubernetes
containers. Thus, we assume that the course build cronjob is run
in a container that has one version of Sphinx installed.
There are two containers, one for the new and one for the old Sphinx
version. The cron.sh code skips over courses whose Sphinx version
does not match the version installed in the container.