Show checkbox question hints below the question by default
In questionnaires, show feedback right under a selected checkbox only
if the new field configuration "checkbox_feedback" is enabled. Many teachers
do not want the feedback/hint to be shown right under the checkbox.
By default, all feedback/hints are shown in the bottom of the question like
it used to be before the summer 2020
(commit b20aee313452a1ac8cb56083070f44deaa8a26d7).

Ensure that feedback is shown for correct answer too
(including the inverse "not selected" feedback).

Fix the old prefix hint feature. If multiple hints start with the same prefix,
only the longest hint is shown. This works when the checkbox hints are shown
in the bottom (the default).

Fixes #115

Show the "multiple choices are selectable" hint only when the answer selected
one checkbox and the correct answer should include multiple.