Use only short language codes for translations
A+ sends the lang parameter to the MOOC-Grader in the grading
protocol. Usually, the language code is in the short format "en"
or "fi", but in some cases in the long format "en-gb".

Courses are used to defining multilingual course contents with
short language codes "en" and "fi". A long language code "en-gb"
fails to find the English translation and the MOOC-Grader may
return Finnish contents if the course default language is Finnish.

This is easy to fix by ignoring the region in the long language
code format. Currently, we do not have a use case for translating
courses differently to both American English and British English,
thus it is not necessary to support long language codes.
If it becomes necessary, it would be possible to support long
language codes: e.g., for lang "en-gb", we could check
the existence of translations in the order "en-gb", "en", "fi"
instead of "en-gb", "fi" (skipping the possibility of the short
format "en").