Fix the model solutions of radio button questions
In questionnaires, the model solutions of radio button questions had
a bug when the question contained both correct and neutral choices.
The field initial value contained multiple choices, but that is invalid
for radio buttons. Model solutions are rendered by setting the correct
answers to the form initial values. Therefore, model solutions had
invalid data. The HTML output showed validation errors.
Loading the model solution crashed the mooc-grader if the question
included feedback because the field cleaned_data becomes None when
the input is invalid. The None value causes crashes in
the compare_values method.

Fixes #99