Handle errors in included exercise config files
When the exercise config.yaml uses the "include" field, errors in
parsing the file were previously not handled and they caused
crashes.

In addition, fix a bug in loading the included config template.
When the exercise config.yaml used "include" without
"template_context", there was a bug in loading the included
template. It was not loaded because yaml.safe_load does not
accept a file path (string) as a parameter. Either a file object
or a string of YAML data to parse should be passed.