put configure files on root level again
a few options are dependent on the relative locations of the files
so we cannot add an intermediary directory yet. Would have to figure out
which options need to be modified and whether there are other consequences
for changing the file locations.

Let's just hope that there is no name collision