Support Scala3 REPL in the ignorerepl modifier of text questions
In text questions of questionnaires, the `ignorerepl` modifier
means that students are allowed to answer entire lines of text
copy-pasted from the Scala REPL even though the question just
asks for a single element from that line.

REPL output in Scala2 looks like this:
`res0: Int = 123456`
while in Scala3, it is like this:
`val res0: Int = 123456`

This commit adds support for the starting `val `.
It may still also be omitted.

Fixes #134