Enable LTI parameters for grading
If a MOOC-Grader exercise is configured to enable LTI,
then the A+ front sends LTI parameters about the student in
addition to the normal grader protocol parameters.
LTI parameters include, e.g., the student's name, email and student id.

When a submission is graded, the grading container can read
the LTI parameters from the file `lti.json` in the working directory
inside the container.

Note: this commit does not add any real LTI protocol support to
the MOOC-Grader. At this stage, this is only a workaround to
pass the student's name and email to the grader via the LTI
protocol parameters.

The MOOC-Grader LTI exercise must also set
`lti_aplus_get_and_post: True` in its config.yaml.
Since the MOOC-Grader does not support the real LTI protocol,
A+ must use the A+ grading protocol to connect to the grader.
That is enabled via the setting.