Fix group_errors grading in questionnaires
Grading question groups with the `group_errors` setting was broken.
This commit fixes it so that no points are granted if some question
in the group is answered incorrectly.
In addition, no "correct/incorrect" symbols are shown for individual
questions, but they are shown for the whole question group.

Fix #149