Fix crash in questionnaire grading due to recent changes
Fix the bug in the commit 57ec3416996736215608731b1c586e369f83edc9.
The questionnaire grading crashed if the questionnaire did not use
the group_errors setting because the form class's instance variable
`self.group_errors` only existed when the setting was enabled.
The instance variable is now always initialized with `False`.