Move generated personalized exercises to course git repo
The personalized exercise instance directories are moved into
the course git repo so that teachers may create and modify them
via git as any other content in the course. The script for
generating personalized exercise instances is also modified and
moved to the course git repository for the teachers to use.

Fixes apluslms#138